Check for description for changes
Web api has method to check network availability, is not yet
implemented.

get long in itemdatasource class was missing.

Changed order table ids to long.

Pairing fragment more robust, checks table existence first.

+++ Selfie/app/src/main/java/database/ItemDataSource.java
@@ -451,7 +451,7 @@ public class ItemDataSource {
                 cursor.getLong(cursor.getColumnIndex(SelfieDatabase.KEY_ITEM_ID)),
                 cursor.getString(cursor.getColumnIndex(SelfieDatabase.KEY_ITEM_NAME)),
                 cursor.getFloat((cursor.getColumnIndex(SelfieDatabase.KEY_PRICE))),
-                cursor.getInt((cursor.getColumnIndex(SelfieDatabase.KEY_CATEGORY_ID))),
+                cursor.getLong((cursor.getColumnIndex(SelfieDatabase.KEY_CATEGORY_ID))),
                 cursor.getInt((cursor.getColumnIndex(SelfieDatabase.KEY_LIKES))),
                 (IsActive==1)?true:false,
                 cursor.getInt((cursor.getColumnIndex(SelfieDatabase.KEY_CALORIES))),
